ml-service/requirements.txt
flask==2.3.2
pillow==10.0.0
pytesseract==0.3.10
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
joblib==1.3.2
opencv-python-headless==4.8.1.78
gunicorn==20.1.0